refactor: rename loop/utils to groupNodeUtils
