+++ frontend/src/components/nodes/loops/groupNodeUtils.ts
